Attempt to download the binary using cookies and scraping the URL
The download URL is auto-generated on each visit and is only valid in
combination with the Cookie that was sent with the page during the URL
generation. Visit the main page to collect the needed Cookie and scrape the download URL from it. Then use that for the install function.